backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
django-filter>=23.5
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
celery>=5.3
redis>=5.0
channels>=4.0
channels_redis>=4.1
daphne>=4.0
django-environ>=0.11
gunicorn>=21.2
Pillow>=10.2
qrcode>=7.4
reportlab>=4.0
weasyprint>=61.0
django-celery-beat>=2.5
